fix some compile problems
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2692 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/demo/dirent.h
@@ -45,7 +45,7 @@ struct DIR
 	struct dirent		dd_dir;
 
 	/** _findnext handle */
-	intptr_t			dd_handle;
+	int64_t			dd_handle;
 
 	/**
          * Status of search:
